Adding guard around default migration copy

+++ Exec.ps1
@@ -268,7 +268,9 @@ SMTP_DOCKER=$($SmtpDocker)
 		}
 	}
 
-	Copy-Item -Path "migrations/db.$DbEngine" "migrations/db" -Recurse
+	if (!(Test-Path -Path "migrations/db.$DbEngine")) {
+		Copy-Item -Path "migrations/db.$DbEngine" "migrations/db" -Recurse
+	}
 
 	docker exec -it $WebContainer composer update
